refactor: 이미지 처리 및 바코드 스캔 로직 CouponProcessor로 분리
